requirements.txt
Flask==2.3.2
gunicorn==21.2.0
crewai==0.28.8
crewai-tools==0.1.7
python-dotenv==1.0.1
requests==2.31.0
jinja2==3.1.2
Werkzeug==2.3.7
